# documents/robo-factory-swagger.yaml
swagger: "2.0"
info:
  title: "ROBO-Factory"
  description: "ROBO-Factory API Documentation"
  version: "1.0.0"
host: "api.robo-factory.io"
basePath: /v1
tags:
- name: "robots"
  description: "Everything about robots"
  externalDocs:
    description: "Find out more"
    url: "http://robo-factory.io"
schemes:
- "http"
paths:
  /robots:
    post:
      tags:
      - "robots"
      description: "Create a robot" 
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: body
        name: createRobot
        description: The robot to create.
        required: true
        schema:
          $ref: "#/definitions/Robots"
      responses:  
        '201':
          description: "Created"
          schema:
            $ref: "#/definitions/Robots"
        "400":
          description: "Bad Request"
          schema:
            $ref: "#/definitions/Error4xx"
    get:
      tags:
      - "robots"    
      description: "Retreive all robots"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      responses:  
        '200':
          description: "Ok"
          schema:
            type: "array"
            items:
              type: "object"
              properties:
                name :
                  type: "string"
                imageURL:
                  type: "string"
        "400":
          description: "Bad Request"
          schema:
            $ref: "#/definitions/Error4xx"

  /robots/filterbytype:
    get:
      tags:
      - "robots"    
      description: "Retreive all robots filtered by type"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "type"
        in: "query"
        description: "Robot type values for filter"
        type: "string"
        required: true
        items:
          type: "string"
          enum:
          - "good"
          - "bad"
        collectionFormat: "multi"
      responses:  
        '200':
          description: "Ok"
          schema:
            type: "array"
            items:
              type: "object"
              properties:
                name :
                  type: "string"
                imageURL:
                  type: "string"
        "400":
          description: "Bad Request"
          schema:
            $ref: "#/definitions/Error4xx"
  /robots/{robotName}:
    get:
      tags:
      - "robots"    
      description: "Retreive a specific robot"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: path
        name: robotName
        description: name of robot to retreive.
        required: true
        type: integer
      responses:  
        '200':
          description: "Ok"
          schema:
            $ref: "#/definitions/Robots"
        "400":
          description: "Bad Request"
          schema:
            $ref: "#/definitions/Error4xx"
        "404":
          description: "Not Found"
          schema:
            $ref: "#/definitions/Error4xx"
    put:
      tags:
      - "robots"
      description: "Update a robot"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: path
        name: robotName
        description: name of robot to update.
        required: true
        type: integer
      - in: body
        name: createRobot
        description: The robot to create.
        required: true
        schema:
          type: "object"
          required:
            - "type"
            - "cost"
            - "deliveryTime"
            - "imageURL"
          properties:
            type :
              type: "string"
              enum: 
              - "good"
              - "bad"
            cost :
              type: "integer"
            deliveryTime :
              type: "integer"
            imageURL:
              type: "string"
      responses:  
        '200':
          description: "Ok"
          schema:
            $ref: "#/definitions/Robots"
        "400":
          description: "Bad Request"
          schema:
            $ref: "#/definitions/Error4xx"
        "404":
          description: "Not Found"
          schema:
            $ref: "#/definitions/Error4xx"
    delete:
      tags:
      - "robots"
      description: "Delete a robot"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: path
        name: robotName
        description: name of robot to retreive.
        required: true
        type: integer
      responses:  
        '204':
          description: "Deleted"
        "400":
          description: "Bad Request"
          schema:
            $ref: "#/definitions/Error4xx"
        "404":
          description: "Not Found"
          schema:
            $ref: "#/definitions/Error4xx"


definitions:
  Robots:
    type: "object"
    required:
      - "name"
      - "type"
      - "cost"
      - "deliveryTime"
      - "imageURL"
    properties:
      name:
        type: "string"
      type :
        type: "string"
        enum: 
        - "good"
        - "bad"        
      cost :
        type: "integer"
      deliveryTime :
        type: "integer"
      imageURL:
        type: "string"
  Error4xx:
    type: "object"
    properties:
      error:
        type: "string"